Code/rpi_deployment/requirements.txt
# Required Python packages for Raspberry Pi deployment
ultralytics>=8.0.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=1.11.0
torchvision>=0.12.0
RPi.GPIO>=0.7.0